requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

matplotlib==3.10.0
numpy==2.1.3
openai-whisper==20250625
pandas==2.2.3
scipy==1.15.3
sounddevice==0.5.5
streamlit==1.45.1
thefuzz==0.22.1
